tab tweaks, output refactor
