+++ src/components/organisms/AssetActions/Trade/Output.module.css
@@ -1,0 +1,21 @@
+.output {
+  border-top: 1px solid var(--border-color);
+  padding-top: var(--spacer);
+  padding-bottom: var(--spacer);
+  display: grid;
+  gap: var(--spacer);
+  grid-template-columns: 1fr 1fr;
+}
+
+.output p {
+  font-weight: var(--font-weight-bold);
+  margin-bottom: calc(var(--spacer) / 8);
+}
+
+.output [class*='token'] {
+  white-space: normal;
+}
+
+.output [class*='token'] > figure {
+  display: none;
+}
